Fixed errors at level A and B
Inconsistent indexing/ naming of Level A and B tasks lead to numerous errors that needed to be fixed.  Most inconsistencies related to whether task numbers should be unique across all tasks or only w/n each level.  New policy is that all task numbers are unique. This change affects how a given task is indexed within the list of tasks for each level.