Add the CPMD_PER_UNIT overhead value as measured on the MC^2 machine
This is obtained from a measurement program which randomly loops through
a 512KB memory allocation. The loop time with a hot cache is recorded
and compared to the loop time with a cold cache. This experiment was
repeated 1000 times. Obvious outliers were removed and the average and
maximum were computed and applied to Level-C and Level-B respectively.

These values work out to an average cache line load time of 29ns, which
seems not-obviously incorrect.